rf: move ``NormalizeMotionParams`` interface into ``confounds``
