+++ niworkflows/interfaces/confounds.py
@@ -1,13 +1,14 @@
 # emacs: -*- mode: python; py-indent-offset: 4; indent-tabs-mode: nil -*-
 # vi: set ft=python sts=4 ts=4 sw=4 et:
 """Select terms for a confound model, and compute any requisite expansions."""
-
+import os
 import re
 import numpy as np
 import pandas as pd
 from functools import reduce
 from collections import deque, OrderedDict
 from nipype.utils.filemanip import fname_presuffix
+from nipype.utils.misc import normalize_mc_params
 from nipype.interfaces.base import (
     traits,
     TraitedSpec,
@@ -16,6 +17,33 @@ from nipype.interfaces.base import (
     isdefined,
     SimpleInterface,
 )
+
+
+class _NormalizeMotionParamsInputSpec(BaseInterfaceInputSpec):
+    in_file = File(exists=True, mandatory=True, desc="the input parameters file")
+    format = traits.Enum(
+        "FSL", "AFNI", "FSFAST", "NIPY", usedefault=True, desc="output format"
+    )
+
+
+class _NormalizeMotionParamsOutputSpec(TraitedSpec):
+    out_file = File(exists=True, desc="written file path")
+
+
+class NormalizeMotionParams(SimpleInterface):
+    """Convert input motion parameters into the designated convention."""
+
+    input_spec = _NormalizeMotionParamsInputSpec
+    output_spec = _NormalizeMotionParamsOutputSpec
+
+    def _run_interface(self, runtime):
+        mpars = np.loadtxt(self.inputs.in_file)  # mpars is N_t x 6
+        mpars = np.apply_along_axis(
+            func1d=normalize_mc_params, axis=1, arr=mpars, source=self.inputs.format
+        )
+        self._results["out_file"] = os.path.join(runtime.cwd, "motion_params.txt")
+        np.savetxt(self._results["out_file"], mpars)
+        return runtime
 
 
 class _ExpandModelInputSpec(BaseInterfaceInputSpec):
