+++ niworkflows/interfaces/utils.py
@@ -12,7 +12,6 @@ from collections import OrderedDict
 import scipy.ndimage as nd
 from nipype import logging
 from nipype.utils.filemanip import fname_presuffix
-from nipype.utils.misc import normalize_mc_params
 from nipype.interfaces.base import (
     traits,
     isdefined,
@@ -26,36 +25,6 @@ from .. import __version__
 
 
 LOG = logging.getLogger("nipype.interface")
-
-
-class _NormalizeMotionParamsInputSpec(BaseInterfaceInputSpec):
-    in_file = File(exists=True, mandatory=True, desc="the input parameters file")
-    format = traits.Enum(
-        "FSL", "AFNI", "FSFAST", "NIPY", usedefault=True, desc="output format"
-    )
-
-
-class _NormalizeMotionParamsOutputSpec(TraitedSpec):
-    out_file = File(exists=True, desc="written file path")
-
-
-class NormalizeMotionParams(SimpleInterface):
-    """
-    Convert input motion parameters into the designated convention.
-
-    """
-
-    input_spec = _NormalizeMotionParamsInputSpec
-    output_spec = _NormalizeMotionParamsOutputSpec
-
-    def _run_interface(self, runtime):
-        mpars = np.loadtxt(self.inputs.in_file)  # mpars is N_t x 6
-        mpars = np.apply_along_axis(
-            func1d=normalize_mc_params, axis=1, arr=mpars, source=self.inputs.format
-        )
-        self._results["out_file"] = os.path.join(runtime.cwd, "motion_params.txt")
-        np.savetxt(self._results["out_file"], mpars)
-        return runtime
 
 
 class _GenerateSamplingReferenceInputSpec(BaseInterfaceInputSpec):
